+++ plant_tracker/routes/plants.py
@@ -11,7 +11,12 @@ from flask import (
 from pathlib import Path
 
 from plant_tracker.core.utils import default_if_prop_none
-from plant_tracker.core.geodata import get_all_geodata
+from plant_tracker.core.geodata import (
+    GeodataPoint,
+    GeodataPolygon,
+    process_gdata_and_assign_location,
+    get_all_geodata
+)
 from plant_tracker.forms.add_image import (
     AddImageForm,
     get_image_data_from_form,
@@ -87,30 +92,19 @@ def edit_plant(plant_id: int = None):
         elif request.method == 'POST':
             plant = get_plant_data_from_form(session=session, form_data=request.form, plant_id=plant_id)
             if form['geodata'].data:
-                loc_name = f'{plant.species.common_name}#{plant_id}'
-                is_polygon = request.form["shape_type"] == 'polygon'
                 gtype = GeodataType(f'plant_{request.form["shape_type"]}')
-                gdata = request.form['geodata']
-                if plant.plant_location and plant.plant_location.geodata_key:
-                    # Existing data
-                    if gdata != plant.plant_location.geodata.data:
-                        # Update shape data
-                        plant.plant_location.geodata.data = gdata
-                    if gtype != plant.plant_location.geodata.geodata_type:
-                        # Log change in shape
-                        plant.plant_location.geodata.geodata_type = gtype
-                        plant.plant_location.geodata.is_polygon = is_polygon
-                else:
-                    # Create new plant location
-                    plant.plant_location = TablePlantLocation(
-                        plant_location_name=loc_name,
-                    )
-                    plant.plant_location.geodata = TableGeodata(
-                        geodata_type=gtype,
-                        name=loc_name,
-                        is_polygon=is_polygon,
-                        data=gdata
-                    )
+
+                if plant.plant_location is None:
+                    plant.plant_location = TablePlantLocation()
+
+                plant.plant_location = process_gdata_and_assign_location(
+                    session=session,
+                    table_obj=plant.plant_location,
+                    form_data=request.form,
+                    geo_type=gtype,
+                    alt_name=f'{plant.species.common_name}#{plant_id}'
+                )
+
             session.add(plant)
             flash(f'Plant {plant.species.scientific_name} ({plant.plant_id}) successfully updated', 'success')
             return redirect(url_for('plant.get_all_plants'))
@@ -185,8 +179,8 @@ def get_all_plants(species_id: int = None):
                 pt.species.scientific_name,
                 pt.species.common_name,
                 default_if_prop_none(pt, 'plant_source'),
-                default_if_prop_none(pt, 'region.region_name'),
-                default_if_prop_none(pt, 'sub_region.sub_region_name'),
+                default_if_prop_none(pt, 'plant_location.region.region_name'),
+                default_if_prop_none(pt, 'plant_location.sub_region.sub_region_name'),
                 [
                     {'url': url_for('plant.edit_plant', plant_id=pt_id),
                      'icon': 'bi-pencil', 'val_class': 'icon edit me-1'},